# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_LanreAkintayo_outpost_internal_dto.CreateEndpointRequest:
    properties:
      description:
        maxLength: 500
        type: string
      rate_limit:
        maximum: 1000
        minimum: 1
        type: integer
      recipient_id:
        maxLength: 255
        type: string
      url:
        type: string
    required:
    - url
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.DeliveryAttemptSummary:
    properties:
      attempt_number:
        type: integer
      endpoint_id:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_models.DeliveryStatus'
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.EndpointResponse:
    properties:
      application_id:
        type: string
      consecutive_failures:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      rate_limit:
        type: integer
      recipient_id:
        type: string
      secret:
        type: string
      status:
        $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_models.EndpointStatus'
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.EventResponse:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      event_type_id:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      payload:
        items:
          type: integer
        type: array
      recipient_id:
        type: string
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.IngestEventResponse:
    properties:
      created_at:
        type: string
      event_type:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      queued_deliveries:
        type: integer
      recipient_id:
        type: string
      status:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.ReplayEventRequest:
    properties:
      failed_only:
        type: boolean
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.ReplayEventResponse:
    properties:
      attempts:
        items:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.DeliveryAttemptSummary'
        type: array
      event_id:
        type: string
      queued_deliveries:
        type: integer
    type: object
  github_com_LanreAkintayo_outpost_internal_dto.SendEventRequest:
    properties:
      event_type:
        maxLength: 255
        minLength: 1
        type: string
      idempotency_key:
        type: string
      payload:
        items:
          type: integer
        type: array
      recipient_id:
        type: string
    required:
    - event_type
    - payload
    type: object
...
  github_com_LanreAkintayo_outpost_internal_dto.UpdateEndpointRequest:
    properties:
      description:
        type: string
      rate_limit:
        maximum: 1000
        minimum: 1
        type: integer
      recipient_id:
        type: string
      status:
        $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_models.EndpointStatus'
      url:
        type: string
    type: object
  github_com_LanreAkintayo_outpost_internal_models.DeliveryStatus:
    enum:
    - pending
    - processing
    - delivered
    - failed
    - dead_letter
    type: string
    x-enum-varnames:
    - DeliveryStatusPending
    - DeliveryStatusProcessing
    - DeliveryStatusDelivered
    - DeliveryStatusFailed
    - DeliveryStatusDeadLetter
  github_com_LanreAkintayo_outpost_internal_models.EndpointStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - EndpointStatusActive
    - EndpointStatusInactive
  github_com_LanreAkintayo_outpost_internal_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    name: Outpost Support
  description: High-performance, fault-tolerant webhook delivery platform with exponential
    retries, rate limiting, HMAC signing, and dead-letter queues.
  license:
    name: MIT
  title: Outpost Webhook Delivery Engine API
  version: "1.0"
paths:
...
  /api/v1/endpoints:
    get:
      description: List all registered webhook endpoints for the authenticated application
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EndpointResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List endpoints
      tags:
      - Endpoints
    post:
      consumes:
      - application/json
      description: Register a new webhook destination URL and generate an HMAC signing
        secret (whsec_...)
      parameters:
      - description: Endpoint configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.CreateEndpointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EndpointResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create endpoint
      tags:
      - Endpoints
  /api/v1/endpoints/{id}:
    delete:
      description: Delete a webhook endpoint and remove its subscriptions
      parameters:
      - description: Endpoint UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete endpoint
      tags:
      - Endpoints
    get:
      description: Fetch details of a specific webhook endpoint
      parameters:
      - description: Endpoint UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EndpointResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get endpoint by ID
      tags:
      - Endpoints
    put:
      consumes:
      - application/json
      description: Update URL, description, status (active/inactive), or rate limit
        of an endpoint
      parameters:
      - description: Endpoint UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Endpoint update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.UpdateEndpointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EndpointResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update endpoint
      tags:
      - Endpoints
...
  /api/v1/events:
    post:
      consumes:
      - application/json
      description: Ingest a webhook event, check idempotency, schedule delivery attempts
        to subscribed endpoints, and enqueue tasks
      parameters:
      - description: Event payload and routing parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.SendEventRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.IngestEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ingest event
      tags:
      - Events
  /api/v1/events/{id}:
    get:
      description: Fetch details of an ingested event by UUID
      parameters:
      - description: Event UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.EventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get event by ID
      tags:
      - Events
...
  /api/v1/events/{id}/replay:
    post:
      consumes:
      - application/json
      description: Queue new delivery attempts for an existing event (failed attempts
        only by default)
      parameters:
      - description: Event UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Replay options (failed_only flag)
        in: body
        name: request
        schema:
          $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.ReplayEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_dto.ReplayEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_LanreAkintayo_outpost_internal_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Replay single event
      tags:
      - Events
...
securityDefinitions:
  BearerAuth:
    description: Enter your API key with the Bearer prefix, e.g. 'Bearer op_live_...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"